+++ src/crawlee/beautifulsoup_crawler/beautifulsoup_crawler.py
@@ -74,8 +74,11 @@ class BeautifulSoupCrawler(BasicCrawler[BeautifulSoupCrawlingContext]):
             request=context.request,
             session=context.session,
             proxy_info=context.proxy_info,
-            send_request=context.send_request,
             add_requests=context.add_requests,
+            send_request=context.send_request,
+            get_data=context.get_data,
+            push_data=context.push_data,
+            export_to=context.export_to,
             http_response=result.http_response,
         )
 
@@ -134,9 +137,12 @@ class BeautifulSoupCrawler(BasicCrawler[BeautifulSoupCrawlingContext]):
             request=context.request,
             session=context.session,
             proxy_info=context.proxy_info,
-            send_request=context.send_request,
-            add_requests=context.add_requests,
             enqueue_links=enqueue_links,
+            add_requests=context.add_requests,
+            send_request=context.send_request,
+            get_data=context.get_data,
+            push_data=context.push_data,
+            export_to=context.export_to,
             http_response=context.http_response,
             soup=soup,
         )