+++ src/crawlee/playwright_crawler/playwright_crawler.py
@@ -64,8 +64,11 @@ class PlaywrightCrawler(BasicCrawler[PlaywrightCrawlingContext]):
         yield PlaywrightCrawlingContext(
             request=context.request,
             session=context.session,
-            send_request=context.send_request,
             add_requests=context.add_requests,
+            send_request=context.send_request,
+            get_data=context.get_data,
+            push_data=context.push_data,
+            export_to=context.export_to,
             proxy_info=context.proxy_info,
             page=crawlee_page.page,
         )