+++ tests/unit/basic_crawler/test_context_pipeline.py
@@ -37,6 +37,9 @@ async def test_calls_consumer_without_middleware() -> None:
         add_requests=AsyncMock(),
         session=Session(),
         proxy_info=AsyncMock(),
+        get_data=AsyncMock(),
+        push_data=AsyncMock(),
+        export_to=AsyncMock(),
     )
 
     await pipeline(context, consumer)
@@ -60,6 +63,9 @@ async def test_calls_consumers_and_middlewares() -> None:
             add_requests=AsyncMock(),
             session=context.session,
             proxy_info=AsyncMock(),
+            get_data=AsyncMock(),
+            push_data=AsyncMock(),
+            export_to=AsyncMock(),
         )
         events.append('middleware_a_out')
 
@@ -73,6 +79,9 @@ async def test_calls_consumers_and_middlewares() -> None:
             add_requests=AsyncMock(),
             session=context.session,
             proxy_info=AsyncMock(),
+            get_data=AsyncMock(),
+            push_data=AsyncMock(),
+            export_to=AsyncMock(),
         )
         events.append('middleware_b_out')
 
@@ -84,6 +93,9 @@ async def test_calls_consumers_and_middlewares() -> None:
         add_requests=AsyncMock(),
         session=Session(),
         proxy_info=AsyncMock(),
+        get_data=AsyncMock(),
+        push_data=AsyncMock(),
+        export_to=AsyncMock(),
     )
     await pipeline(context, consumer)
 
@@ -106,6 +118,9 @@ async def test_wraps_consumer_errors() -> None:
         add_requests=AsyncMock(),
         session=Session(),
         proxy_info=AsyncMock(),
+        get_data=AsyncMock(),
+        push_data=AsyncMock(),
+        export_to=AsyncMock(),
     )
 
     with pytest.raises(RequestHandlerError):
@@ -131,6 +146,9 @@ async def test_handles_exceptions_in_middleware_initialization() -> None:
         add_requests=AsyncMock(),
         session=Session(),
         proxy_info=AsyncMock(),
+        get_data=AsyncMock(),
+        push_data=AsyncMock(),
+        export_to=AsyncMock(),
     )
 
     with pytest.raises(ContextPipelineInitializationError):
@@ -159,6 +177,9 @@ async def test_handles_exceptions_in_middleware_finalization() -> None:
         add_requests=AsyncMock(),
         session=Session(),
         proxy_info=AsyncMock(),
+        get_data=AsyncMock(),
+        push_data=AsyncMock(),
+        export_to=AsyncMock(),
     )
 
     with pytest.raises(ContextPipelineFinalizationError):